improvment: consistent error toasting on all request
