+++ frontend/src/core/network/requests-static.ts
